move model related stuff to its own file + move final_check_eh to the proper file
